# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff/login:
    post:
      operationId: staffLogin
      tags: [staff]
      summary: Log in with staff credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/accounts:
    get:
      operationId: getStaffAccounts
      tags: [staff]
      summary: List all staff accounts (admin only)
      security:
        - staffToken: []
      responses:
        "200":
          description: List of staff accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffAccount"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createStaffAccount
      tags: [staff]
      summary: Create a new staff account (admin only)
      security:
        - staffToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffAccountRequest"
      responses:
        "201":
          description: Staff account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccount"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/sessions/revoke-all:
    post:
      operationId: revokeAllStaffSessions
      tags: [staff]
      summary: Revoke all active staff sessions (admin only)
      description: >
        Sets a server-side not-before timestamp so all tokens issued before
        this moment are immediately rejected. Staff currently logged in will
        be signed out on their next API request.
      security:
        - staffToken: []
      responses:
        "200":
          description: Sessions revoked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeSessionsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/accounts/{id}/revoke-sessions:
    post:
      operationId: revokeStaffAccountSessions
      tags: [staff]
      summary: Revoke sessions for a single staff account (admin only)
      description: >
        Sets a per-account not-before timestamp so all tokens issued before
        this moment for that specific account are immediately rejected.
        Only the targeted staff member is signed out; other staff are unaffected.
      security:
        - staffToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sessions revoked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeSessionsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/security-events:
    get:
      operationId: getSecurityEvents
      tags: [staff]
      summary: List security audit log events (admin only)
      description: >
        Returns a paginated list of security events such as session revocations,
        ordered most-recent first.
      security:
        - staffToken: []
      responses:
        "200":
          description: Security events returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityEvent"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/accounts/{id}/deactivate:
    patch:
      operationId: deactivateStaffAccount
      tags: [staff]
      summary: Deactivate a staff account (admin only)
      security:
        - staffToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccount"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/accounts/{id}/activate:
    patch:
      operationId: activateStaffAccount
      tags: [staff]
      summary: Reactivate a staff account (admin only)
      security:
        - staffToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccount"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/me:
    get:
      operationId: getMyStaffProfile
      tags: [staff]
      summary: Get the current staff member's own profile (includes email)
      security:
        - staffToken: []
      responses:
        "200":
          description: Own profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccount"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/{id}/email:
    patch:
      operationId: updateStaffEmail
      tags: [staff]
      summary: Set or update a staff member's email address (canonical path)
      security:
        - staffToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffEmailRequest"
      responses:
        "200":
          description: Email updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccount"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/accounts/{id}/email:
    patch:
      operationId: updateStaffEmailAlias
      tags: [staff]
      summary: Set or update a staff member's email address (alias)
      security:
        - staffToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffEmailRequest"
      responses:
        "200":
          description: Email updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccount"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    staffToken:
      type: http
      scheme: bearer
  schemas:
...
    StaffLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    StaffSession:
      type: object
      properties:
        token:
          type: string
        staffId:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, housekeeper]
      required:
        - token
        - staffId
        - username
        - displayName
        - role
    StaffAccount:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, housekeeper]
        active:
          type: boolean
        email:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - username
        - displayName
        - role
        - active
        - createdAt
    UpdateStaffEmailRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
      required:
        - email
...
    CreateStaffAccountRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, housekeeper]
      required:
        - username
        - password
        - displayName
...
    RevokeSessionsResponse:
      type: object
      properties:
        revokedBefore:
          type: string
          description: ISO timestamp — all tokens issued before this moment are now rejected
      required:
        - revokedBefore
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error