# lib/api-zod/src/generated/types/staffAccount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StaffAccountRole } from "./staffAccountRole";

export interface StaffAccount {
  id: number;
  username: string;
  displayName: string;
  role: StaffAccountRole;
  active: boolean;
  email?: string | null;
  createdAt: string;
}